removed pkg variable which required package.json
The package.json was required in this package as `pkg` but  that variable was never called in the file.  As a result, I got an error when trying to use this package in a project w/ webpack. (I tried just using the json loader for webpack of course, but that didn't seem to solve the problem).  As a result I ended up having to remove this require from my project wondered if it might be better not to include it at all if its not used.